Don't repost after message deletion
Privileged users can use `delete` to delete a message, but SmokeDetector
would still repost a question if it would be bumped. If a message gets
deleted now, the post ID and site name are stored in a file and before
posting, SmokeDetector checks whether a message reporting that post has
been deleted once.