Added `ignore` command for privileged users
If a post is not good enough to mark it as false positive and not bad
enough to mark it as true positive, but you still don't want to get more
alerts about it, reply `ignore` (without backticks). Then it won't be registered as tp/fp, but SmokeDetector won't post alerts about it anymore. You need to be a privileged user to execute this command.