Fix bug when ignoring post
When ignoring a post, we fetch a tuple `(post_id, site)` via
`fetch_post_id_and_site_from_msg_content()`. `fetch_...` was recently
changed to return `(post_id, site, type_of_post)`. Therefore, we were
adding this to the ignorelist, and when checking the ignorelist, we
compared it with `(post_id, site)`. Of course, the comparison returned
`False`.