Escape ] in titles
Currently, we don't escape ] (link formatting) when posting the title, so it generates weird markdown results like http://chat.stackexchange.com/transcript/message/19703439#19703439. It seems like we only need to escape ], so this should work. We should probably do this in usernames & commits, but I'm not too sure there's many examples of this in the wild (except for this commit. ;P)