new rule: blacklisted websites on non-fitness
We didn't have any rule for the _body_ of posts on non-fitness/health/biology sites.   
I added such a rule, with the regex for site domains including "workout", "fitness", or "diet".

Also changed .com to .co  in domain patterns because .co.uk appears now.   
And added a couple of specific domains.