Soften 'tp' for answer message
If you tp an answer, it still gets recorded as a true positive in metasmoke. Typically, people know what they're doing in the distinction of tp/tpu, so we shouldn't make this an 'error' message. So, this commit softens the message and makes it an 'informative' message -- it only replies if the quiet action syntax isn't used.