Quota rollover is lost if it happens when Smokey is down.
If Smokey has crashed or rebooted when the API rollover happens, Smokey doesn't detect it and we lose all sense of what the lowest quota remaining might be.
This change effectively adds an extra line soon after the customary start line that helps mitigate that loss.

The gold-standard way to track quota levels, is to update the quota in a file. But, this file would see about 10K read/write cycles a day.  Do we want to do that?