Add some protection for EOFErrors
On the most troublesome files, if an EOFError is raised, we will delete
the file and then re-raise the exception. This should make Smokey able
to fix any error it comes across. I don't believe any of the other files
have given us trouble, but if they do, I can add this for them too.