Add DeletionWatcher
At the moment, this will only operate for the Tavern and for questions.
This has not been unit tested (hypocrite, right?) so it may break stuff.

Short summary: each time Smokey posts a report, a watcher will be
created for 2 minutes (the deletion period) to check if the post has
been deleted. If the post is deleted, Smokey will delete the message
posted in the Tavern. If nothing occurs in two minutes, the thread
should just exit.