Check if a post is being re-reported with manual reports
In some cases, posts are acidentally manually reported when it was
already reported several posts before. This check will make sure that
double reports don't occur. has_already_been_posted contains a memory of
about 15 posts -- if the post is in this memory, and the post was not
marked as a false positive (because this might be correcting that
mistake), the post will not be reported again.