Change classmethod to staticmethod, remove suppression.  --autopull
This change fixes the yelling about 'self' vs. 'cls', but also saves us about 10 bytes.  Since we don't refer to `self.something` in either of these methods, we can run them as static methods, and not have to worry about passing an instance of the class itself to the method.  (This is working in a DIRTY run of Smokey currently)