gitmanager: lowercase compare for GH error msgs.  --autopull
This way, we don't have to worry about mistyping the upper or lowercase characters incorrectly, and we can explicitly only worry about content.  I don't trust *not* making a mistake there with case of character, so let's do the lowercase comparison instead.

Also, got annoyed about the wavy lines on DeletionWatcher's class methods, so add a suppression for those warnings and inspections in IDEs.  It was so minor a change it didn't warrant its own commit.