Implement two separate Flake8 tests
First, we are going to define two separate `tox` configuration files, one for code, one for tests.  The difference is in what is excluded, and what we ignore.

Secondly, we're modifying the Travis and Circle CI procedures to have an extra `flake8` test.  The first thing we do is run Flake8 tests against the `./test/*` directory, to make sure all of it passes Flake8 restrictions (minus the line length enforcement we have in place for the SmokeDetector code).  Since we have dedicated configuration files for each Flake8 run, one for the tests and one for the code, we need to have two separate runs on two separate directory locations, with two different configuration files.

 Note that this renames `tox.ini` to `tox_code.ini`, which means we will have to specify the configs to use if we adjust the CI processes.