Add try/except for ws.py for debugger breakpoint definition for ValueErrors.
This is necessary for our brief run of Smokey from within a PyCharm debugger so we can add catchable breakpoints that will halt the processes *before* passing things to our Exception Handler.  This test-run will run sometime in the next few days, but when not running in a Debugger, it will raise the exception it gets either way because it raises the same error one way or another.