Enable ability to go into 'Attach Debugger' mode w/ PyCharm debuggers
First, we introduce a new `Debugging` class in `debugging.py`; this handles determining whether or not we need to launch the Debugger, based on environment variables, and constructs different ways of getting the arguments in ways you can manipulate what's being launched later (a la `ws.py`)

Secondly, we modify `nocrash.py` a little.  We make it have a 'copy' of the current environment variables already made available to it, from launch.  We then go and check, via the import of the `Debugging` class, whether or not we're enabling attaching to the debugger.  If `Debugging` is enabled, then add values to the copied environment variables set, defining a secondary set of environment variables that we will ultimately pass down to `ws.py`.

Thirdly, we modify `ws.py`.  This is where the most important modifications would happen.  Given that `nocrash.py` spawns `ws.py`, when using a debugger such as that in PyCharm it is impossible to attach to `nocrash.py` when it's already running and also track activity in `ws.py`.  This makes it impossible to 'break' when you want it to (such as an unhandled exception in `ws.py`) to get an idea of what was happening, where, and what triggered a given error.  (It also makes it impossible to get a memory snapshot of what variables have what values, etc.)  In this case, since we'll be using the same debugging 'server', we have `ws.py` run its own checks on the environment passed down to it, via the `Debugging` class import, and determine early on in the code whether we're attaching to a debugger or not.  This check happens before we even begin to launch `ws.py`, and prints to the command line that it's attaching the debugger.  Failure to import or find the corresponding package or module will trigger a failure case and totally exit entirely via exit code 6; failure to attach to debugger fails and exits via the same exit code.

There's a downside here though.  We have to include a copy of PyCharm's `pycharm-debug.egg` package in the local tree.  This adds a little bit to the files, but is mandatory if we want to be able to include 'debugging' capabilities we can easily switch on and off.  So, fourth item of business is to make sure that `pycharm-debug.egg` is properly included in the repository, which means making a "do not ignore" entry in `.gitignore` for that file.  (This also adds a comment a few lines earlier about the .DS_Store file.)