findspam.py: New reason "Bad fragment in link"
For the time being, it specifically matches URLs ending with "-reviews"
(with a couple of oddball variations from old spam).  This is a common
MO for the Indian pharma spammer.  Most of these spams are detected just
fine, but they are mutating brand names etc. incessantly and might one
day succeed in getting past the usual filters.  This is added as just
one more layer of "defense in depth".

It should be relatively easy to see how to add more fragments to the
rule in the future.

There are two new test cases to check both branches of the regex at
the heart of this rule.

Current Metasmoke report for a similar query suggests 3494/3494 TP rate:
https://metasmoke.erwaysoftware.com/search?utf8=%E2%9C%93&title=&body_is_regex=1&body=%28%5E%7C%5B%5EA-Za-z0-9_%5D%29%5BA-Za-z0-9_%5D%2B-reviews%28-%28canada%7C%28and%7Cor%29-scam%29%29%3F%2F%3F%3C%2Fa%3E&username=&why=&site=&feedback=&reason=&user_rep_direction=%3E%3D&user_reputation=0&commit=Search