don't depend on GlobalVars
Instead, depend on helpers.log() for reporting parsing problems.

This makes the error less visible, but avoids the potential risks of using an exception
(see discussion on PR#668, which this supersedes).

Also coincidentally fix the api_response= handling so that a None value will not
override a default (notably, passing Metasmoke results with "score": null would
fail further down in the code when you attempt to manipulate them).