Remove NobodyNada from PA list, due to errors
With NobodyNada in here, we see this evil error:

```
"{'required': 1, 'users': ['angussidney', 'Aralun', 'ArtOfCode-', 'AWegnerGitHub', 'doppeldeveloper', 'ferrybig', 'Fortunate-MAN', 'Glorfindel83', 'hichris1234', 'honnza', 'magisch', 'Manishearth', 'metasmoke', 'NobodyNada', 'ProgramFOX', 'superplane39', 'teward', 'tripleee', 'Undo1'], 'approve_by_comment': {'approve_regex': '^!!\\/approve', 'enabled': True}, 'author_approval': {'ignored': True}, 'reset_on_push': {'enabled': True}, 'reject_value': -1}" is an invalid value for reviewers: 'NobodyNada' is not a collaborator of this repo
```

Until such time this is fixed in the system backend here on GH to give NobodyNada access as a collaborator on the repo, it's going to continue to error.  This is resulting in automatic rejection on PullApprove.  So, this will 'repair' that.