GlobalVars.parser.unescape(): avoid deprecation warning
HTMLParser.unescape() was supposed to have been removed in Python 3.5
(but was allegedly still forgotten in).  The canonical replacement is simply
html.unescape().  Hack it in to the parser object so the existing code can
continue to work without deprecation warnings (though it's only actually
used in 3 places; we should really avoid using a global for this eventually).