AU should be a special case with queue size 1
AU is by far the most spammed site outside of SO. It was also hit by a troll recently. There is enough spare API calls to use queue size 1 for AU (doubling the number of requests to it). 

Also: since "time sensitive" category means that queue size drops to 1 between UTC 4 and UTC 12, the sites that have queue size 1 already need not be in that category, it's redundant.  

Finally, writers.SE is now writing.SE, the old name still works for the API but will not match the new site name provided by websocket connections.