Attempt to fix the 'user stupidity' risk.
While the risk of "multiple reasons" is small, this fixes the formatting issue(s) that were identified by Art, but only if the user is properly following syntax of `"reason" "another reason"`.  This, however, is *not* what we originally had 'designed' this around; if a user is doing this, they should be given a failure case probably, instead of a "Success".