Better handle 'pickle data was truncated' unpickling failures.
This happened on the teward/Solar Flare instance today, where a pickle got its data truncated somehow. The exact error was that 'pickle data was truncated', and the only solution to fix this issue was to remove the pickle in question.  This should resolve this type of error going forward now by removing the pickle being loaded the same way as we'd handle an EOFError for an incomplete file, but only for truncated pickle data errors.