Increase resistance of blacklist/watchlist commands to MS downtime.
If MS is down we get a connection error problem.  When that happens, we should force code_permissions to be False.  However, to get the proper chat reply that MS is down and a PR made because we can't determine if someone is code privileged or not, we need a metasmoke_down tracker in the 'add to blacklist' function so we can determine which chat reply to send.