Capitalise the first letter of MS auto-posted comments
Sometimes, if someone replies to SD to add a comment, they will not capitalize the first letter of their reply as it reads more cleanly as a sentence (e.g. @SmokeDetector that post looks spammy). However, this causes the comment to look weird on MS as the reply text. This change automatically capitalises the first letter so that it looks more professional on MS.

https://chat.stackexchange.com/transcript/message/48826618#48826618