Tell user why pattern is invalid
Adjusted CmdException message for do_blacklist(), to show what pattern it is processing that triggered the error, for debug reasons. Theoretically, this error should never be triggered, as the command requires one argument and hence must be in the form "!!/blacklist-* some_regex", and hence the split() call will return two items and item [1] can be accessed as usual. The fact that an IndexError is raised indicates that msg.content_source contains no whitespace, which means that no argument is provided (hence there is some bug in other part of the code, likely in chatcommunicate.py)