dnsresolver.py: Pythonic DNS resolver
This is the prerequisite bits for a Pythonic implementation of the `dig` command.  It replaces the prior subprocess calls to `dig` and makes sure to utilize System DNS configurations to configure the DNS resolver.  This will replace the `dig` command with a more secure Python implementation of `dig` by using `dnspython`