Remove coveralls as a separate test
While it's probably a good idea for this to be a
separate CI pass/fail test, it needs to be something that normally
passes for PRs (i.e. it needs to reflect the actual change in
coverage caused by the PR). Currently, because one test
is automatically skipped when testing a PR, the coverage
for PRs is automatically substantially lower and the test
will always fail, unless the PR itself adds significant
test coverage.
-autopull