Revert reversal of blacklists/watchlist.
At a minimum, this may give us data as to the claim that having
the lists reversed is be faster. However, my primary reasons for
doing this are

A) at multiple points in the past, people have made intentionally
adjusted the order of the entries such that some entries don't get
shadowed by other entries. All of these adjustments have been done
under the assumption that the list is added to the regex in the
order the entries are in in the files, not the reverse. Having these
reversed doesn't just make that work ineffective, it makes every
one of those changes accomplish exactly the opposite of what they
intended. Finding all the places where such changes were made is
difficult. We'd need substantial improvement in performance to
justify not just nulling-out those changes, but making them actively
bad.

B) People making changes to the lists in the future are more likely
to assume that the lists are not reversed, rather than that they are.

-autopull