Consistent use of GlobalVars.bot_repo_slug -autopull
We shouldn't store a copy of GlobalVars.bot_repo_slug in the
class, unless that's going to be the primary place that information
is stored.

As it was, the class was using both its local copy that was
only updated when the class was instantiated, and the GlobaVars
copy. The different versions were used in various functions. This
could result in hard to track down bugs, if the GlobalVars
copy was ever changed during operation (it doesn't but...).
Any way we slice it, everything in the class should either use
the copy that was made upon class instantiation, or the copy
that's in GlobalVars, not a mix of both.

Given that the GlobalVars copy is primary, I changed all to use
that copy.