Tentative async approve/reject
This is an attempt to solve #4291 . The observation of SD behaviour is that it randomly loses connection to CHQ after an approve/reject command is issued. I suspect therefore that this is due to the git operation hanging up the entire command listener thread, as SD is still able to post 'Goodbye, cruel world' message to CHQ but not accept any command. This is an attempt to make approve/reject asynchronous, in order to verify the cause of the bug.