DNSPython: 2.0.0 changes (use resolve, not query)
Looks like dnsresolver class with the modifiable resolver class in here for configuration and use in the `!!/ip` call we used a while ago for DNS lookups of things (when we were debugging differing DNS records for domains) did not get the changes that #5117 was going to implement.  So, implement.