bad_keywords: update text-x to reduce FP
The original appears to be generating FP due, primarily, to the
Unicode based `\b` which we use for bookending matching a transition
between `>/` and `</`. An alternative regex:
`test(?<=/test)[\W_]*+x[\W_]*+(?:pro|core|360)` would also match
all the posts we've seen, but would probably miss a couple of matches
within a few posts.
-autopull