Use consistent merge strategy.
We may or may want to only use --ff-only here. However, if we do,
then we should be using it consistently for all of the merges we
perform in the different paths for this overall process.

This just changes it to use the same default set of merge strategies
as we have been using in the past. That doesn't mean that we can't
or shouldn't change that, just that if we do change it, we should
be consistent about it.

It would probably be good for us to take a more thorough look at
how we're handling this overall.

--autopull