Don't output "resolved to" DNS resolutions in CI tests.
When we get an error in these tests, these are a lot of lines of
text to wade through in order to see what's happening. I'm not
sure what the benefit is in having them printed all the time,
but it definitely makes it less convenient from my point of view
to find where the actual error is.

autopull