For notifications, don't fetch the users in the room unless needed.
The prior code was making a request to chat for the current users
in the room regardless of if there were any users to notify for the
room and regardless of if those users had specified to only be
pinged when they are in the room.

This changes the logic to only fetch from chat the list of users
currently in the room when there's at least one user users who
has requested to be notified for reports in that room only
when they are present in the room.