Correct casing of coffee and tea names to fix grammar in replies
Previously, the names in the "COFFEE" list were all capitalized, while the names in the "TEAS" list were all lowercase. Grammatically, in the context of SmokeDetector's reply, lowercase is correct, so all names have been converted to lowercase, except where a particular name is a proper noun, in which case, it is correctly rendered in uppercase.