Scan threads re-check the queue when done with a site.
Testing prior to this change indicated that there are times when the
queue can grow, but that the limits on launching threads,
particularly the CPU limit, would result in times when the queue
grew, but did not get immediately serviced. In fact, entries existed
in the queue for a human noticeable time prior to the vagaries of
the WebSocket, etc. causing more threads to be launched.

Prior attempts at having the scan threads loop didn't work well, but
that was mostly prior to having the CPU limit. Having the queue
rechecked after the thread is complete will result in existing
threads being preferred for new scans (i.e. they are likely to check
just after dropping the CPU load, so are more likely to find the
CPU loading to be below saturation).