For SO, set BodyFetcher queue depth to 2.
This is a quick change to reduce the flood of requests we're making
for SO. Given how often SO is being scanned at the moment,
this should have only a small effect on response time.

Overall, we'll need to take a better look at how scans are being
launched for SO, given its volume and that we are not notified
via the WebSocket of every post.

autopull