GH Actions: Wipe the Python minor version dependent venv.
The primary point of having and caching the venv is to reduce the
time it takes to install the site_packages. Caching the Python minor
version dependent things, including pip, is not desired. This removes
those and allows the `python3 -m venv env` command to regenerate
them while keeping the site_package.

As I recall, keeping the site_packages is even more beneficial for
more recent versions of Python, which include more recent versions
of `pip`. This is the case because `pip` substantially changed
the algorithm it uses to resolve dependencies. The new algorithm
can stall out and take hours, eventually being killed by GitHub
Actions, rather than the few seconds required by earlier
versions of `pip`.

autopull