Reduce regex CPU costs: initial \W or \w not accounting Unicode \b
autopull

The Unicode based `\b` we use on the watchlist and bad keywords,
`(?w:\b)`, can result in *lots* of "word" boundaries within
sequences which would be considered entirely `\W` if ASCII
definitions were used. The main issue is that many of these regexes
were written with the assumption that they would be starting from
a ASCII `\b`. Overall, this results in a few to several orders of
magnitude more CPU resources than is required to have the same
results. The quick an dirty way to fix this is to add a `\b` to the
start of the affected regexes in order to spend very little
processing when it's a `(?w:\b)`, but not a `\b`.