Add a complete second verification pass for NS verification
The intent is to prevent CI testing from failing due to intermittent
DNS resolution issues. While we already do a second pass, it doesn't
seem to resolve the issues we're seeing on particular domains.

In addition, with the prior code we don't have easy access to the
domain that is failing. With this change, the failing domains are
collected into a list. If we continue to have issues with a
specific domain that isn't actually a failure and something we want
to remove from then list, then we can just exclude it from being
reported as an error in CI testing.