Update MIN_TESTED_PYTHON_VERSION = (3, 8, 16)
We haven't tested against 3.7 for a while now.
I'm not updating MIN_PYTHON_VERSION, because it's supposed to
be the version below which we *know* we have problems. I'm not
recalling anything specific which _requires_ 3.7+, but I do
recall looking at some items and choosing to not use the ones
which were backwords compatible with older Python versions, so
it's possible we require 3.7 at this point.

autopull