Revert "Revert "Revert "Merge PR#8552: Replace `datetime.utcnow()` with `datetime.now(tz=timezone.utc)`"""
This reverts commit 7de29485b93492704cfafe07ebf7e0a4378d3f95.

The changes done in the original merge request #8552 introduce hard breakages to SmokeDetector.

Introducing this change makes datetimes timezone-aware, and we also have datetimes coming via
websocket that are NOT tz-aware. This introduced a hard ValueError during SmokeDetector initialization.

This is a hard revert.  DO NOT revert this revert without checking with @teward first!