Tweak broad support/service watch expression to exclude systemd log FPs
Anecdotally, most of the FPs that I've seen coming recently
from this generic watch expression are the systemd-formatted
log/debug output that this tweak will exclude.

Empirically, this takes all-time MS results
from: 1475 TP|376 FP|11 NAA
to  : 1474 TP|352 FP|9 NAA.

Eventually, when the spam wave of endless variations that this
is intended to catch has ended, this entire watch should probably
be removed, but it is far too soon for that now.

autopull